move RestoreOptimizer callback to patch api
